backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pywebview==5.3

# Strands Agents
strands-agents>=0.2.0

# HTTP Client for OpenRouter
httpx>=0.27.0

# WebSocket Support
websockets>=12.0

# MCP (Model Context Protocol) Support
mcp>=1.0.0

# Data Validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
beautifulsoup4>=4.12.0

# Type Extensions
typing-extensions>=4.0.0
sqlalchemy>=2.0.0
